[UI][I] Add boilerplate code for new module selection page
Adds the boilerplate code for the new classes added to re-implement the
module selection page. The newly added class
SelectLocalModuleRepresentationPage will eventually replace
SelectProjectPage.

The new implementation splits the different responsibilities into
different classes to make it easier to modify the logic to allow for
multiple shared module and to make the logic easier to understand.