[FIX][I] #488 Adjust module comparison for multiple open projects
Adjusts the module based comparison used in IntelliJProjectImpl to also
use project information for object comparisons. This is needed as
module names, which were previously used to compare such objects, are
not unique between different projects.

To resolve this problem, the project name and project file base path
were added to the comparison and hash code method. This should ensure
that the object comparison functions correctly between modules belonging
to different projects.

Thereby fixes #488.

Adds the ProjectClosedHandler to end the session when a project
containing a shared module is closed. This is needed as the new
comparison uses the project object. If such a project is closed, we have
to ensure that there is no session running on a module of the project as
the comparison otherwise tries to access the disposed project object,
leading to an exception.