[INTERNAL][I] Show non-shareable modules in contact pop menu
Adjusts ContactPopMenu to also show non-shareable modules. Such modules
are disabled in the menu and are displayed in a separate group.
Furthermore, the carry a tooltip explaining that/why they can't be
shared.

Subsequently removes the user notification when encountering a
non-shareable module.

This re-structuring was done to avoid annoying the user with unnecessary
notifications.